requirements.txt
pandas==2.2.*
numpy==1.26.*
scipy==1.11.*
scikit-learn==1.5.*
imbalanced-learn==0.12.*
xgboost==2.0.*
matplotlib==3.8.*
seaborn==0.13.*
google-cloud-storage==2.*
google-cloud-bigquery==3.*
google-cloud-bigquery-storage==2.*
db-dtypes==1.2.*              
google-cloud-aiplatform==1.*  
fastapi==0.111.*
uvicorn[standard]==0.30.*
jinja2==3.*
joblib==1.4.*
